+++ .gitignore
@@ -56,7 +56,7 @@ playground.xcworkspace
 # you should judge for yourself, the pros and cons are mentioned at:
 # https://guides.cocoapods.org/using/using-cocoapods.html#should-i-check-the-pods-directory-into-source-control
 #
-# Pods/
+Pods/
 #
 # Add this line if you want to avoid checking in source code from the Xcode workspace
 # *.xcworkspace
@@ -92,3 +92,4 @@ fastlane/test_output
 iOSInjectionProject/
 *.xcuserstate
 neutral.love/Managers/APIManager/Secret.swift
+neutral.love/Resources/GoogleService-Info.plist
