+++ .swiftlint.yml
@@ -1,0 +1,12 @@
+excluded:
+  - Pods
+
+line_length:
+   warning: 150
+
+disabled_rules:
+    - trailing_whitespace
+    - identifier_name
+    - nesting
+    - void_function_in_ternary
+    - unused_closure_parameter
