--- neutral.love/.swiftlint.yml
@@ -1,12 +1,0 @@
-excluded:
-  - Pods
-
-line_length:
-   warning: 150
-
-disabled_rules:
-    - trailing_whitespace
-    - identifier_name
-    - nesting
-    - void_function_in_ternary
-    - unused_closure_parameter
